soroban_security_detectors_sdk

Module detector

Detector module This module provides the Detector trait and macros for implementing detectors.

Public members

detector! macro for defining a detector. It automatically creates the structure for the provided type_name in the attribute and implements Detector trait. It can be applied to a single function with type_name attribute and follows check function signature.

detectors! macro for defining multiple detectors at once. It can be applied to a list of functions with type_name attribute and follows check function signature.

Detector trait for implementing a detector. It has a single method check that takes a Codebase and returns an optional vector of DetectorResult.

DetectorResult struct for representing the result of a detector. It contains the file path, start and end offsets, and an optional map of extra information. Extra information is used to store a map of symbol replacements in the detector template. For example, if the detector template contains a symbol $NAME, the extra information can be used to replace it with the actual name.

DetectorReportTemplate trait for implementing a detector report template. It has methods for generating the report title, body, and closing.

CombinedDetector a union trait to force the implementor to implement both Detector and DetectorReportTemplate traits.

SorobanDetector a boxed version of CombinedDetector.

DetectorOpaque a struct that is used to wrap a raw pointer to a detector. It is used to operate with detectors using C API.

Structs

DetectorOpaque

WARNING: This struct is used to wrap a raw pointer to a detector. In you write detectors in a separate library, you should not use this struct to cast the pointer to Detector.

DetectorResult

DetectorResult struct Represents the result of a detector.

Traits

CombinedDetector

CombinedDetector trait A union trait to force a Detector implementation to implement both Detector and DetectorReportTemplate traits.

Detector

Detector trait The base Detector functional interface.

DetectorReportTemplate

DetectorReportTemplate trait The DetectorReportTemplate trait is used to define the report template for a detector. This trait exposes the Detector metadata used for generating the report.

Type Aliases

SealedCodebase

SorobanDetector

SorobanDetector type An alias for a boxed version of CombinedDetector.