soroban_security_detectors_sdk

Macro detector

macro_rules! detector {
    (
        #[type_name = $tname:ident]
        $(#[$attr:meta])*
        $vis:vis fn $name:ident < $($gen:tt),* > ( $($params:tt)* )
        $(-> $ret:ty)?
        $(where $($where:tt)*)?
        $body:block
    ) => { ... };
    () => { ... };
}

Detector macro This macro is used to define a detector. It accepts a function (signature and body) with a type_name attribute. The function signature must follow the check function signature from the Detector trait. It automatically creates the structure for the provided type_name in the attribute and implements the Detector trait. The DetectorReportTemplate trait should be implemented to satisfy the CombinedDetector contract.