soroban_security_detectors_sdk

Struct Codebase

pub struct Codebase<S> { /* private fields */ }

Implementations

impl Codebase<OpenState>

pub fn parse_and_add_file( &mut self, file_path: &str, content: &mut str, ) -> Result<(), SDKErr>

Parse the file and add it to the codebase.

Errors

SDKErr::AddDuplicateItemError If the file is already added.

Panics

Panics if the internal fname_ast_map is None.

pub fn build_api( &mut self, files: &HashMap<String, String>, ) -> Result<Box<Codebase<SealedState>>>

Builds the API from the codebase.

Errors

Returns an error if the user crate root cannot be found or if parsing fails.

Panics

Panics if the internal fname_ast_map is None.

pub fn get_item_by_id(&self, id: u32) -> Option<NodeKind>

impl Codebase<SealedState>

pub fn files(&self) -> impl Iterator<Item = Rc<File>>

pub fn contracts(&self) -> impl Iterator<Item = Rc<Contract>>

pub fn get_node_source_code(&self, node_id: u32) -> Option<String>

pub fn functions(&self) -> impl Iterator<Item = Rc<Function>> + '_

pub fn get_parent_container(&self, id: u32) -> Option<NodeKind>

pub fn find_node_file(&self, id: u32) -> Option<Rc<File>>

pub fn get_children_cmp<F>(&self, id: u32, comparator: F) -> Vec<NodeKind>
where F: Fn(&NodeKind) -> bool,

pub fn get_children_cmp_cast<F, C>(&self, id: u32, comparator: F) -> Vec<C>
where F: Fn(&NodeKind) -> bool, C: From<NodeKind>,

pub fn node_type(&self, node: &NodeKind) -> NodeType

pub fn compare_types(&self, a: &NodeKind, b: &NodeKind) -> bool

pub fn inline_function(&self, func: &Rc<Function>) -> Function

pub fn get_expression_type(&self, node_id: u32) -> NodeType

pub fn get_symbol_type(&self, scope_id: u32, symbol: &str) -> Option<NodeType>

pub fn get_function_by_name( &self, scope_id: u32, name: &str, ) -> Option<Definition>

pub fn lookup_symbol_origin( &self, scope_id: u32, symbol: &str, ) -> Option<NodeKind>